app/requirements.txt
alembic==1.0.3
asn1crypto==0.24.0
Authlib==0.10
awscli==1.16.61
boto3==1.9.50
botocore==1.12.51
braintree==3.49.0
cachetools==3.0.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
cryptography==2.4.2
docutils==0.14
Flask==1.0.2
Flask-Assets==0.12
Flask-Login==0.4.1
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
google-api-python-client==1.7.4
google-auth==1.6.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.2.0
httplib2==0.12.0
idna==2.7
itsdangerous==1.1.0
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.0
oauthlib==2.1.0
pipdeptree==0.13.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
python-dateutil==2.7.5
python-dotenv==0.9.1
python-editor==1.0.3
PyYAML==3.13
requests==2.20.1
requests-oauthlib==1.0.0
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.14
uritemplate==3.0.0
urllib3==1.24.1
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.2.1
